+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/model/Formintakea.java
@@ -189,7 +189,7 @@ public class Formintakea implements Serializable {
 
     private int hashCode = Integer.MIN_VALUE;// primary key
 
-    private Long _id;// fields
+    private Integer _id;// fields
     private String _radioSex;
     private String _effDate;
     private String _cboxSharing;
@@ -308,7 +308,7 @@ public class Formintakea implements Serializable {
     private String _cboxAssistinaddictions;
     private String _cboxAssistinfinance;
     private String _radioActive;
-    private Long _providerNo;
+    private Integer _providerNo;
     private String _cboxEmployment;
     private String _mentalHealthConcerns;
     private String _radioTakemedication;
@@ -332,7 +332,7 @@ public class Formintakea implements Serializable {
     private String _radioOnlinecheck;
     private String _allergicToMedicationName;
     private Date _formCreated;
-    private Long _demographicNo;
+    private Integer _demographicNo;
     private String _cboxGotoemergencyroom;
     private String _frequencyOfSeeingEmergencyRoomDoctor;
     private String _cboxWsib;
@@ -352,7 +352,7 @@ public class Formintakea implements Serializable {
 	/**
 	 * Constructor for primary key
 	 */
-	public Formintakea (Long _id) {
+	public Formintakea (Integer _id) {
 		this.setId(_id);
 		initialize();
 	}
@@ -361,8 +361,8 @@ public class Formintakea implements Serializable {
 	 * Constructor for required fields
 	 */
 	public Formintakea (
-		Long _id,
-		Long _demographicNo) {
+			Integer _id,
+			Integer _demographicNo) {
 
 		this.setId(_id);
 		this.setDemographicNo(_demographicNo);
@@ -385,7 +385,7 @@ public class Formintakea implements Serializable {
 *  generator-class="native"
 *  column="ID"
 */
-    public Long getId () {
+    public Integer getId () {
         return _id;
     }
 
@@ -393,7 +393,7 @@ public class Formintakea implements Serializable {
 	 * Set the unique identifier of this class
      * @param _id the new ID
      */
-    public void setId (Long _id) {
+    public void setId (Integer _id) {
         this._id = _id;
         this.hashCode = Integer.MIN_VALUE;
     }
@@ -2171,7 +2171,7 @@ public class Formintakea implements Serializable {
     /**
 	 * Return the value associated with the column: provider_no
      */
-    public Long getProviderNo () {
+    public Integer getProviderNo () {
         return _providerNo;
     }
 
@@ -2179,7 +2179,7 @@ public class Formintakea implements Serializable {
 	 * Set the value related to the column: provider_no
      * @param _providerNo the provider_no value
      */
-    public void setProviderNo (Long _providerNo) {
+    public void setProviderNo (Integer _providerNo) {
         this._providerNo = _providerNo;
     }
 
@@ -2531,7 +2531,7 @@ public class Formintakea implements Serializable {
     /**
 	 * Return the value associated with the column: demographic_no
      */
-    public Long getDemographicNo () {
+    public Integer getDemographicNo () {
         return _demographicNo;
     }
 
@@ -2539,7 +2539,7 @@ public class Formintakea implements Serializable {
 	 * Set the value related to the column: demographic_no
      * @param _demographicNo the demographic_no value
      */
-    public void setDemographicNo (Long _demographicNo) {
+    public void setDemographicNo (Integer _demographicNo) {
         this._demographicNo = _demographicNo;
     }
